Fix bug when specifying to order_asc
This can never have worked. For now, going to keep the current behaviour
of only adding the order_asc parameter if we explicitly do want to order
asc, and never add order_asc=false (so relying on the API's default of
false). In theory just always adding the parameter should make no
difference and would be cleaner, but as it would affect the request sent
I don't want to do that in a patch version. Eventually, should probably
make this a nullable bool and send nothing by default, but allowing the
user to specify whether the order_asc parameter with either value, but
not going to do that now.